Support pattern for field to be absent in response
Add a a *A pattern, which can be used in the response messages to make sure the field is absent in the response. There are times when a field  should not be echoed back or set in the response. *A will check if the field is present in the response and if it does it will fail the testcase. The optional *O does not provide as much control as needed.