Add support for environment values on quartz when attribute
The quartz module support the configuration of the jobs with a xml, this
xml normally define the class and a cron expression, this commit add
support for the new Environment property resolver that use various
sources to resolve a given property.

This commit also add test to check that this change doesn't break the
previous usage.

Signed-off-by: Arturo Volpe <avolpe@fintech.works>