feat: Allow write-perm attribute in 00_qi
Now we can specify a write-perm for a view as well as for
a specific field. If the user does not have the specified
write-perm for the view the canAdd/canRemove/canEdit will
return false (unless overridden). If the user does not
have the specified write-perm for a field, it will act as
read-only field.