fix(Issue #208): Jpos hangs on htp response 204 no content
Jpos hangs when a server response is 204 no content. in the case where
the responseBodyOnError is set true and the response status is 204 jpos
hangs becuase there is no response body to process. This causes the
EntityUtils to attempt to process the response, calling the
result.getEntity() which does nothing. This change allows for the
situation where 204 is the response staus, and the responseBodyOnError
to be true but not call the EntityUtils when there is no response body
to process, by adding the status to the context using the mapping value
of HTTP_RESPONSE_STATUS