feat: Add support for '.' on qrest path parameter
This commit adds support for path parameter that has dots ('.').

The dot is a commonly used path parameter, for example, all the URLs with
filenames have a dot for the extension.

I can't find any reference to support dots in this specification:

    https://www.rfc-editor.org/rfc/rfc6570

But in this specification:

    https://www.w3.org/Addressing/URL/url-spec.txt

When describing a
[BNF](https://en.wikipedia.org/wiki/Backus%E2%80%93Naur_form), the dot
is in the 'safe' rule, which is part of the 'segment' rule that defines
the 'path' of an URL.

Other use cases are:

* API Versioning: add the API version in the path, this is already
  partially supported with the wildcard (v2/*)
* minigl code: the mini account codes are a tree of all the parent
  accounts codes separated with a dot.

Signed-off-by: Arturo Volpe <avolpe@fintech.works>

Signed-off-by: Arturo Volpe <avolpe@fintech.works>