Removed an unused SSH reference mistakenly flagged as a vulnerability.
Text files containing literals that reference a package version number
do not pose a security problem if they are not actually used as a
dependency. Such references are harmless and do not impact the security
of the project.

QSAs can verify the actual dependencies used by the jPOS project by
examining the MANIFEST.MF file of the top-level JAR. Only the
dependencies listed there are part of the classloader. Any references
not included in the MANIFEST.MF are not loaded and thus do not affect
the project.