Update VariableFieldPackager.java
The value needs to be set with an empty string if the user did not set it specifically and was null. This mimics unpack where value is set to "" when only a delimiter is found. Due to this not being set in pack, the hexdump called on the msg after pack, results in a NPE.