spin off txn-lite from txn
Just a few classes in the 'txn' module require dbsupport (Open, Close,
DBDebug). In order to use 'txn' as a dependency without carrying the
whole dbsupport set of dependecies, we are creating a new 'txn-lite'
one. txn depends on txn-lite, so despite the extra dependency, this
change shouldn't break existing code.